FIX: changed the deleted and toplvl flags of lfsl to uint32_t (was crashing on opteron, probably due to aligment optimizations)

+++ src/skiplist/skiplist.h
@@ -55,12 +55,12 @@ typedef ALIGNED(CACHE_LINE_SIZE) struct sl_node
 {
   skey_t key;
   sval_t val;
-  intptr_t deleted;
-  int toplevel;
+  uint32_t deleted;
+  uint32_t toplevel;
   struct sl_node* next[1];
 #if defined(DO_PAD)
   uint8_t padding[64 - 32];
-#endif
+#endif 
 } sl_node_t;
 
 typedef ALIGNED(CACHE_LINE_SIZE) struct sl_intset
